Remove lazy workaround
I had added a little hacky fix to get the xml2lua dependency
installed locally and on CI. For some reason lazy wasn't installing it
on its own. I have since found out that this is because I didn't publish
a non-dev version of this package to luarocks, and something about how
dependencies are looked up needed that to be the case. Now that there's
a package published, xml2lua is installed correctly.