update mini.test to get ignore_attr
the attr field of the screenshot varies on different versions of nvim.
if i want to test those in the test matrix, I have to ignore it. It
might be possible to configure them all the same, but it sounds like a
headache to get it right. Really what I care about is the text output
anwyay.