Add integration test (#25)
This loads a bun test in a child nvim process, runs tests, and asserts that the output is as expected. This is a super high level test that goes through the full testing flow. I think it'll be helpful to have a few of these in addition to the lower level unit tests.